update cart button for mobile
