+++ hamza-client/src/modules/layout/templates/nav/menu-mobile/connect-wallet/connect-button.tsx
@@ -22,12 +22,7 @@ import {
 import { MdOutlineWallet } from 'react-icons/md';
 import { useSwitchNetwork } from 'wagmi';
 
-//Todo: If chain unsupported?
-export const WalletConnectButton = ({
-    cart: cartState,
-}: {
-    cart?: Omit<Cart, 'beforeInsert' | 'afterLoad'> | null;
-}) => {
+export const WalletConnectButton = () => {
     const { error, isLoading, pendingChainId, switchNetwork } =
         useSwitchNetwork();
 
@@ -42,6 +37,8 @@ export const WalletConnectButton = ({
             {({
                 account,
                 chain,
+                openAccountModal,
+                openChainModal,
                 openConnectModal,
                 authenticationStatus,
                 mounted,
@@ -54,17 +51,11 @@ export const WalletConnectButton = ({
                     (!authenticationStatus ||
                         authenticationStatus === 'authenticated');
 
-                const totalItems =
-                    cartState?.items?.reduce((acc: any, item: any) => {
-                        return acc + item.quantity;
-                    }, 0) || 0;
-
                 return (
                     <div
                         {...(!ready && {
                             'aria-hidden': true,
                             style: {
-                                opacity: 0,
                                 pointerEvents: 'none',
                                 userSelect: 'none',
                             },
@@ -181,58 +172,7 @@ export const WalletConnectButton = ({
                                     flexDirection={'row'}
                                     alignItems={'center'}
                                 >
-                                    <Flex alignSelf={'center'} ml="1rem">
-                                        <LocalizedClientLink href="/cart">
-                                            <Flex
-                                                position="relative"
-                                                width={'100%'}
-                                                color="white"
-                                                _hover={{
-                                                    '.cart-text, .cart-icon': {
-                                                        color: 'primary.green.900',
-                                                    },
-                                                }}
-                                            >
-                                                <Flex
-                                                    flexDirection={'row'}
-                                                    alignSelf={'center'}
-                                                    color={'white'}
-                                                    _hover={{
-                                                        '.cart-icon': {
-                                                            color: 'primary.green.900',
-                                                            transition:
-                                                                'color 0.3s ease-in-out',
-                                                        },
-                                                    }}
-                                                >
-                                                    <HiOutlineShoppingCart
-                                                        className="cart-icon"
-                                                        size={'25px'}
-                                                    />
-                                                </Flex>
-                                                {totalItems > 0 && (
-                                                    <Flex
-                                                        position="absolute"
-                                                        top="-4px"
-                                                        right="-4px"
-                                                        width="15px"
-                                                        height="15px"
-                                                        borderRadius="full"
-                                                        backgroundColor="#EB4C60"
-                                                        justifyContent="center"
-                                                        alignItems="center"
-                                                        fontSize="10px"
-                                                        color="white"
-                                                        fontWeight="700"
-                                                    >
-                                                        <Text fontSize={'10px'}>
-                                                            {totalItems}
-                                                        </Text>
-                                                    </Flex>
-                                                )}
-                                            </Flex>
-                                        </LocalizedClientLink>
-                                    </Flex>
+                                    <AccountMenu />
                                 </Flex>
                             );
                         })()}
